Moved where the custom CRecipientFilter class lives
